requirements/base.txt
Django==1.9.6
Pillow==3.2.0
PyJWT==1.4.0
PyYAML==3.11
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
jsonpickle==0.9.3
sqlparse==0.1.19
